# adapter/src/main/java/eu/cokeman/velomarker/openapi/api/ElevationApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.22.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package eu.cokeman.velomarker.openapi.api;

import eu.cokeman.velomarker.openapi.model.ElevationRequestDto;
import eu.cokeman.velomarker.openapi.model.ElevationResponseDto;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.22.0")
@Validated
@Controller
@Tag(name = "Elevation", description = "the Elevation API")
public interface ElevationApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_CALCULATE_ELEVATION = "/route/elevation";
    /**
     * POST /route/elevation : Calculate elevation profile for a LineString (Copernicus DEM)
     * Samples Copernicus DEM GLO-30 along the provided coordinates and returns elevation profile + gain/loss. Coordinates beyond max-samples are downsampled. 
     *
     * @param elevationRequestDto  (required)
     * @return Elevation profile (status code 200)
     *         or Bad request (invalid coordinates) (status code 400)
     *         or Unauthorized (status code 401)
     *         or DEM source error (status code 502)
     */
    @Operation(
        operationId = "calculateElevation",
        summary = "Calculate elevation profile for a LineString (Copernicus DEM)",
        description = "Samples Copernicus DEM GLO-30 along the provided coordinates and returns elevation profile + gain/loss. Coordinates beyond max-samples are downsampled. ",
        tags = { "Elevation" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Elevation profile", content = {
                @Content(mediaType = "application/json", schema = @Schema(implementation = ElevationResponseDto.class))
            }),
            @ApiResponse(responseCode = "400", description = "Bad request (invalid coordinates)"),
            @ApiResponse(responseCode = "401", description = "Unauthorized"),
            @ApiResponse(responseCode = "502", description = "DEM source error")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = ElevationApi.PATH_CALCULATE_ELEVATION,
        produces = { "application/json" },
        consumes = { "application/json" }
    )
    default ResponseEntity<ElevationResponseDto> calculateElevation(
        @Parameter(name = "ElevationRequestDto", description = "", required = true) @Valid @RequestBody ElevationRequestDto elevationRequestDto
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "{ \"minEleM\" : 5, \"profile\" : [ [ 0.8008281904610115, 0.8008281904610115 ], [ 0.8008281904610115, 0.8008281904610115 ] ], \"gainM\" : 6, \"maxEleM\" : 5, \"lossM\" : 1 }";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}