# adapter/src/main/java/eu/cokeman/velomarker/openapi/api/StatusApi.java
/*
 * NOTE: This class is auto generated by OpenAPI Generator (https://openapi-generator.tech) (7.22.0).
 * https://openapi-generator.tech
 * Do not edit the class manually.
 */
package eu.cokeman.velomarker.openapi.api;

import eu.cokeman.velomarker.openapi.model.LogSourceResponseDto;
import org.springframework.lang.Nullable;
import eu.cokeman.velomarker.openapi.model.WorkerStatusResponseDto;
import io.swagger.v3.oas.annotations.ExternalDocumentation;
import io.swagger.v3.oas.annotations.Operation;
import io.swagger.v3.oas.annotations.Parameter;
import io.swagger.v3.oas.annotations.Parameters;
import io.swagger.v3.oas.annotations.media.ArraySchema;
import io.swagger.v3.oas.annotations.media.Content;
import io.swagger.v3.oas.annotations.media.Schema;
import io.swagger.v3.oas.annotations.responses.ApiResponse;
import io.swagger.v3.oas.annotations.security.SecurityRequirement;
import io.swagger.v3.oas.annotations.tags.Tag;
import io.swagger.v3.oas.annotations.enums.ParameterIn;
import io.swagger.v3.oas.annotations.media.ExampleObject;
import org.springframework.http.HttpStatus;
import org.springframework.http.MediaType;
import org.springframework.http.ResponseEntity;
import org.springframework.validation.annotation.Validated;
import org.springframework.stereotype.Controller;
import org.springframework.web.bind.annotation.*;
import org.springframework.web.context.request.NativeWebRequest;
import org.springframework.web.multipart.MultipartFile;

import jakarta.validation.Valid;
import jakarta.validation.constraints.*;
import java.util.List;
import java.util.Map;
import java.util.Optional;
import jakarta.annotation.Generated;

@Generated(value = "org.openapitools.codegen.languages.SpringCodegen", comments = "Generator version: 7.22.0")
@Validated
@Controller
@Tag(name = "Status", description = "the Status API")
public interface StatusApi {

    default Optional<NativeWebRequest> getRequest() {
        return Optional.empty();
    }

    String PATH_GET_BROUTER_STATUS = "/admin/route/status";
    /**
     * GET /admin/route/status : BRouter worker status (proxies brouter container&#39;s control-api)
     *
     * @return Per-worker state (status code 200)
     */
    @Operation(
        operationId = "getBrouterStatus",
        summary = "BRouter worker status (proxies brouter container's control-api)",
        tags = { "Status" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Per-worker state", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = WorkerStatusResponseDto.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = StatusApi.PATH_GET_BROUTER_STATUS,
        produces = { "application/json" }
    )
    default ResponseEntity<List<WorkerStatusResponseDto>> getBrouterStatus(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"name\" : \"worker-0\", \"pid\" : 0, \"state\" : \"RUNNING\", \"uptime\" : \"uptime\" }, { \"name\" : \"worker-0\", \"pid\" : 0, \"state\" : \"RUNNING\", \"uptime\" : \"uptime\" } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_LIST_LOGS = "/admin/route/logs";
    /**
     * GET /admin/route/logs : List available BRouter log sources
     *
     * @return Available log sources (status code 200)
     */
    @Operation(
        operationId = "listLogs",
        summary = "List available BRouter log sources",
        tags = { "Status" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Available log sources", content = {
                @Content(mediaType = "application/json", array = @ArraySchema(schema = @Schema(implementation = LogSourceResponseDto.class)))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = StatusApi.PATH_LIST_LOGS,
        produces = { "application/json" }
    )
    default ResponseEntity<List<LogSourceResponseDto>> listLogs(
        
    ) {
        getRequest().ifPresent(request -> {
            for (MediaType mediaType: MediaType.parseMediaTypes(request.getHeader("Accept"))) {
                if (mediaType.isCompatibleWith(MediaType.valueOf("application/json"))) {
                    String exampleString = "[ { \"path\" : \"path\", \"modifiedAt\" : 6, \"name\" : \"name\", \"sizeBytes\" : 0 }, { \"path\" : \"path\", \"modifiedAt\" : 6, \"name\" : \"name\", \"sizeBytes\" : 0 } ]";
                    ApiUtil.setExampleResponse(request, "application/json", exampleString);
                    break;
                }
            }
        });
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_RESTART_WORKER = "/admin/route/workers/{index}/restart";
    /**
     * POST /admin/route/workers/{index}/restart : Restart a single BRouter worker
     *
     * @param index  (required)
     * @return Triggered (status code 204)
     */
    @Operation(
        operationId = "restartWorker",
        summary = "Restart a single BRouter worker",
        tags = { "Status" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Triggered")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = StatusApi.PATH_RESTART_WORKER
    )
    default ResponseEntity<Void> restartWorker(
        @Min(value = 0) @Parameter(name = "index", description = "", required = true, in = ParameterIn.PATH) @PathVariable("index") Integer index
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_ROLLING_RESTART = "/admin/route/rolling-restart";
    /**
     * POST /admin/route/rolling-restart : Rolling restart of all BRouter workers
     *
     * @return Triggered (status code 204)
     */
    @Operation(
        operationId = "rollingRestart",
        summary = "Rolling restart of all BRouter workers",
        tags = { "Status" },
        responses = {
            @ApiResponse(responseCode = "204", description = "Triggered")
        }
    )
    @RequestMapping(
        method = RequestMethod.POST,
        value = StatusApi.PATH_ROLLING_RESTART
    )
    default ResponseEntity<Void> rollingRestart(
        
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }


    String PATH_TAIL_LOG = "/admin/route/logs/{name}";
    /**
     * GET /admin/route/logs/{name} : Tail the last N lines of a log file (plain text)
     *
     * @param name Log source name from /admin/route/logs (e.g. worker-0, nginx-access) (required)
     * @param lines  (optional, default to 200)
     * @return Tail content (status code 200)
     */
    @Operation(
        operationId = "tailLog",
        summary = "Tail the last N lines of a log file (plain text)",
        tags = { "Status" },
        responses = {
            @ApiResponse(responseCode = "200", description = "Tail content", content = {
                @Content(mediaType = "text/plain", schema = @Schema(implementation = String.class))
            })
        }
    )
    @RequestMapping(
        method = RequestMethod.GET,
        value = StatusApi.PATH_TAIL_LOG,
        produces = { "text/plain" }
    )
    default ResponseEntity<String> tailLog(
        @Parameter(name = "name", description = "Log source name from /admin/route/logs (e.g. worker-0, nginx-access)", required = true, in = ParameterIn.PATH) @PathVariable("name") String name,
        @Min(value = 1) @Max(value = 5000) @Parameter(name = "lines", description = "", in = ParameterIn.QUERY) @Valid @RequestParam(value = "lines", required = false, defaultValue = "200") Integer lines
    ) {
        return new ResponseEntity<>(HttpStatus.NOT_IMPLEMENTED);

    }

}