# adapter/src/main/resources/api.yaml
openapi: 3.0.3
info:
  title: Route Service API
  version: 1.0.0
  description: |
    Route planning: brouter proxy (semaphore-protected), draft persistence,
    elevation profile (Copernicus DEM). Plus admin endpoints to manage BRouter
    segments (.rd5) and profiles.
servers:
  - url: http://localhost:8085
    description: Local development server
paths:
...
  /route/elevation:
    post:
      summary: Calculate elevation profile for a LineString (Copernicus DEM)
      description: |
        Samples Copernicus DEM GLO-30 along the provided coordinates and returns
        elevation profile + gain/loss. Coordinates beyond max-samples are downsampled.
      operationId: calculateElevation
      tags: [Elevation]
      security: []
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: '#/components/schemas/ElevationRequest'
      responses:
        '200':
          description: Elevation profile
          content:
            application/json:
              schema:
                $ref: '#/components/schemas/ElevationResponse'
        '400':
          description: Bad request (invalid coordinates)
        '401':
          description: Unauthorized
        '502':
          description: DEM source error
...
  /admin/route/logs/{name}:
    get:
      summary: Tail the last N lines of a log file (plain text)
      operationId: tailLog
      tags: [Status]
      parameters:
        - name: name
          in: path
          required: true
          description: Log source name from /admin/route/logs (e.g. worker-0, nginx-access)
          schema:
            type: string
        - name: lines
          in: query
          required: false
          schema:
            type: integer
            default: 200
            minimum: 1
            maximum: 5000
      responses:
        '200':
          description: Tail content
          content:
            text/plain:
              schema:
                type: string
  /admin/route/workers/{index}/restart:
    post:
      summary: Restart a single BRouter worker
      operationId: restartWorker
      tags: [Status]
      parameters:
        - name: index
          in: path
          required: true
          schema:
            type: integer
            minimum: 0
      responses:
        '204':
          description: Triggered
...
components:
  schemas:
...
    ElevationRequest:
      type: object
      required: [coordinates]
      properties:
        coordinates:
          type: array
          minItems: 2
          items:
            type: array
            minItems: 2
            maxItems: 2
            items:
              type: number
              format: double
          description: Ordered list of [lng, lat] pairs along the route
    ElevationResponse:
      type: object
      required: [profile, gainM, lossM, minEleM, maxEleM]
      properties:
        profile:
          type: array
          description: Sampled elevation profile, each entry [distanceFromStartMeters, elevationMeters]
          items:
            type: array
            minItems: 2
            maxItems: 2
            items:
              type: number
              format: double
        gainM:
          type: integer
          description: Total elevation gain in meters
        lossM:
          type: integer
          description: Total elevation loss in meters
        minEleM:
          type: integer
        maxEleM:
          type: integer
...